Message inbox repository code review & update admin panel.

+++ src/Controller/AdminController.php
@@ -2,22 +2,33 @@
 
 namespace App\Controller;
 
+use App\Entity\Announcement;
 use App\Entity\Group;
 use App\Form\GroupForm;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 
 class AdminController extends AbstractController {
-	public function panel(Request $request) {
+	
+	public function content(Request $request) {
+		$this->denyAccessUnlessGranted('ROLE_ADMIN');
+		$em = $this->getDoctrine()->getManager();
 		
+		$blocks = $em->getRepository(Announcement::class)->findAll();
+		
+		return $this->render('admin/block.html.twig', ['blocks' => $blocks]);
+	}
+	
+	public function group(Request $request) {
+		$this->denyAccessUnlessGranted('ROLE_ADMIN');
 		$em = $this->getDoctrine()->getManager();
 		
 		//creating the new group.
 		$group = new Group();
-		$groupForm = $this->createForm(GroupForm::class, $group);
-		$groupForm->handleRequest($request);
-		if($groupForm->isSubmitted() && $groupForm->isValid()) {
-			$group = $groupForm->getData();
+		$form = $this->createForm(GroupForm::class, $group);
+		$form->handleRequest($request);
+		if($form->isSubmitted() && $form->isValid()) {
+			$group = $form->getData();
 			$em->persist($group);
 			$em->flush();
 			
@@ -25,19 +36,21 @@ class AdminController extends AbstractController {
 		
 		$groups = $em->getRepository(Group::class)->findAll();
 		
-		return $this->render('admin/panel.html.twig', ['groups' => $groups, 'groupForm' => $groupForm->createView()]);
+		return $this->render('admin/group.html.twig', ['groups' => $groups, 'form' => $form->createView()]);
 	}
 	
-	public function delete_group($group) {
+	public function group_delete($group) {
+		$this->denyAccessUnlessGranted('ROLE_ADMIN');
 		
 		if(count($group->getChildren()) > 0) {
-			//return
+			return $this->render('admin/group_delete.html.twig', ['message' => "Group" . $group->getName() . " can't be removed because it contains children."]);
 		}
 		
 		$em = $this->getDoctrine()->getManager();
 		$em->remove($group);
 		$em->flush();
 		
-		return $this->render('admin/panel.html.twig', ['groups' => $groups, 'groupForm' => $groupForm->createView()]);
+		return $this->render('admin/group_delete.html.twig', ['message' => "Group" . $group->getName() . " Removed."]);
+		
 	}
 }
